Add summary table generation and file export feature
